pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentum"
version = "1.0.0"
description = "A Python-native framework for building, orchestrating, and observing production-ready AI agent systems."
readme = "README.md"
authors = [{ name = "Agentum Team", email = "team@agentum.dev" }]
license = { file = "LICENSE" }
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["ai", "multi-agent", "workflow", "orchestration", "llm", "agents", "rag", "vector-search"]

dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3.4",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.1",
    "langchain-openai>=0.1",
    "langgraph>=0.0.47",
    "rich>=13.7",
    "typer>=0.12",
    "python-dotenv>=1.1.1",
    "google-cloud-texttospeech>=2.16.2",
    "langchain-google-community>=1.0.4",
    "langchain-google-genai>=2.1.12",
    "redis>=5.0.0",
    "langchain-community>=0.2.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.7.0",
    "langchain>=0.3",
    "graphviz>=0.21",
    "tavily-python>=0.3.0",
    "jinja2>=3.0",
]

[project.urls]
Homepage = "https://github.com/agentum-framework/agentum"
Documentation = "https://agentum.dev"
Repository = "https://github.com/agentum-framework/agentum"
Issues = "https://github.com/agentum-framework/agentum/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
    "black>=24.0",
    "pre-commit>=3.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests that require API keys",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "build>=1.3.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.2",
    "twine>=6.2.0",
]

[project.scripts]
agentum = "agentum.cli.main:app"
